Separated sensors into their own class and added functionality for checking volume using weight

+++ getBeer_sfml/src/hx711.cpp
@@ -1,6 +1,5 @@
 #include "../include/hx711.h"
-
-HX711::HX711(uint8_t pin_dout, uint8_t pin_slk)
+HX711::HX711(unsigned int pin_dout, unsigned int pin_slk)
 : _pin_dout(pin_dout)
 , _pin_slk(pin_slk)
 {
